FIX passing param meter by url instead of body

+++ controller/admin/removeOfficer.controller.js
@@ -9,24 +9,29 @@ import jwtconfig from '../../config/jwtconfig';
 async function RemoveOfficerController(req, res) {
   const header = req.headers['authorization']
   const token = Handler.getTokenFormHeader(header) // Get token from authorization
+  const email = req.params.email
 
   jwt.verify(token, jwtconfig.secret, async function(err, decoded) {
-    const {
-      email
-    } = req.body
 
-    const deleteRow = await Officer.destroy({
-      where: {
-        email: email
-      }})
+    if (email != undefined && email != null) {
+      const deleteRow = await Officer.destroy({
+        where: {
+          email: email
+        }})
 
-    if (deleteRow == 1) { // If found email and deleted
-      console.log('Deleted');
-      res.status(200).json({ success: true })
+      if (deleteRow == 1) { // If found email and deleted
+        console.log('Deleted');
+        res.status(200).json({ success: true })
+      } else {
+        console.log('Cannot delete row');
+        res.status(401).json({ success: false })
+      }
+
     } else {
-      console.log('Cannot delete row');
-      res.status(401).json({ success: false })
+
+      res.status(400).json({ success: false })
     }
+
   });
 
 }
